refactor: Create paths and math packages
- Make geometry package only contain Pose and Bézier curve instead of all math
- Move Path, PathChain, and PathBuilder to a paths package
- Move Vector and Matrix from geometry and MathFunctions from util to math package
- Update all usages to reflect the new packages

BREAKING CHANGE: Package names are now different for math, geometry, and paths
