+++ src/main/java/com/pedropathing/follower/old/Follower.java
@@ -46,12 +46,12 @@ import com.pedropathing.localization.Localizer;
 import com.pedropathing.geometry.Pose;
 import com.pedropathing.localization.PoseUpdater;
 import com.pedropathing.geometry.BezierPoint;
-import com.pedropathing.util.MathFunctions;
-import com.pedropathing.geometry.Path;
-import com.pedropathing.geometry.PathBuilder;
+import math.MathFunctions;
+import com.pedropathing.paths.Path;
+import com.pedropathing.paths.PathBuilder;
 import com.pedropathing.util.PathCallback;
-import com.pedropathing.geometry.PathChain;
-import com.pedropathing.geometry.Vector;
+import com.pedropathing.paths.PathChain;
+import math.Vector;
 import com.pedropathing.util.DashboardPoseTracker;
 import com.pedropathing.util.Drawing;
 import com.pedropathing.control.FilteredPIDFController;
