+++ src/main/java/com/pedropathing/follower/old/FollowerConstants.java
@@ -4,9 +4,9 @@ package com.pedropathing.follower.old;
 import com.acmerobotics.dashboard.config.Config;
 import com.pedropathing.localization.Localizers;
 import com.pedropathing.geometry.Pose;
-import com.pedropathing.util.MathFunctions;
+import math.MathFunctions;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
-import com.pedropathing.geometry.Vector;
+import math.Vector;
 import com.pedropathing.control.CustomFilteredPIDFCoefficients;
 import com.pedropathing.control.CustomPIDFCoefficients;
 import com.pedropathing.control.KalmanFilterParameters;
