+++ src/main/java/com/pedropathing/localization/GoBildaPinpointDriver.java
@@ -30,7 +30,7 @@ import android.os.Build;
 import androidx.annotation.RequiresApi;
 
 import com.pedropathing.geometry.Pose;
-import com.pedropathing.util.MathFunctions;
+import math.MathFunctions;
 import com.qualcomm.hardware.lynx.LynxI2cDeviceSynch;
 import com.qualcomm.hardware.lynx.LynxNackException;
 import com.qualcomm.robotcore.hardware.I2cAddr;
