+++ src/main/java/com/pedropathing/localization/Localizer.java
@@ -3,7 +3,7 @@ package com.pedropathing.localization;
 import com.pedropathing.geometry.Pose;
 import com.qualcomm.robotcore.hardware.IMU;
 
-import com.pedropathing.geometry.Vector;
+import math.Vector;
 
 /**
  * This is the Localizer class. It is an abstract superclass of all localizers used in Pedro Pathing,
