+++ src/main/java/com/pedropathing/localization/localizers/OTOSLocalizer.java
@@ -5,8 +5,8 @@ import com.qualcomm.robotcore.hardware.HardwareMap;
 
 import com.pedropathing.localization.Localizer;
 import com.pedropathing.geometry.Pose;
-import com.pedropathing.util.MathFunctions;
-import com.pedropathing.geometry.Vector;
+import math.MathFunctions;
+import math.Vector;
 import static com.pedropathing.localization.constants.OTOSConstants.*;
 
 /**
