+++ src/main/java/com/pedropathing/localization/localizers/ThreeWheelIMULocalizer.java
@@ -10,10 +10,10 @@ import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
 
 import com.pedropathing.localization.Encoder;
 import com.pedropathing.localization.Localizer;
-import com.pedropathing.geometry.Matrix;
+import math.Matrix;
 import com.pedropathing.geometry.Pose;
-import com.pedropathing.util.MathFunctions;
-import com.pedropathing.geometry.Vector;
+import math.MathFunctions;
+import math.Vector;
 import com.pedropathing.util.NanoTimer;
 
 /**
