+++ src/main/java/com/pedropathing/localization/localizers/ThreeWheelLocalizer.java
@@ -7,10 +7,10 @@ import static com.pedropathing.localization.constants.ThreeWheelConstants.*;
 
 import com.pedropathing.localization.Encoder;
 import com.pedropathing.localization.Localizer;
-import com.pedropathing.geometry.Matrix;
+import math.Matrix;
 import com.pedropathing.geometry.Pose;
-import com.pedropathing.util.MathFunctions;
-import com.pedropathing.geometry.Vector;
+import math.MathFunctions;
+import math.Vector;
 import com.pedropathing.util.NanoTimer;
 
 /**
